LPS-63095 Move portal-metatype to portal-configuration-metatype
