+++ modules/portal-configuration/portal-configuration-metatype/bnd.bnd
@@ -1,9 +1,9 @@
 Bundle-Name: Liferay Portal Metatype
-Bundle-SymbolicName: com.liferay.portal.metatype
+Bundle-SymbolicName: com.liferay.portal.configuration.metatype
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.portal.metatype.annotations,\
-	com.liferay.portal.metatype.definitions
+	com.liferay.portal.configuration.metatype.annotations,\
+	com.liferay.portal.configuration.metatype.definitions
 Include-Resource: @../../../lib/portal/bnd.jar!/aQute/bnd/annotation/metatype/*
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Metatype Ext
