+++ modules/portal-configuration/portal-configuration-metatype/build.gradle
