+++ modules/portal-configuration/portal-configuration-metatype/src/main/java/com/liferay/portal/configuration/metatype/definitions/ExtendedObjectClassDefinition.java
@@ -12,17 +12,23 @@
  * details.
  */
 
-package com.liferay.portal.metatype.definitions;
+package com.liferay.portal.configuration.metatype.definitions;
 
-import org.osgi.framework.Bundle;
-import org.osgi.service.metatype.MetaTypeService;
+import java.util.Map;
+import java.util.Set;
+
+import org.osgi.service.metatype.ObjectClassDefinition;
 
 /**
  * @author Iván Zaera
  */
-public interface ExtendedMetaTypeService extends MetaTypeService {
+public interface ExtendedObjectClassDefinition extends ObjectClassDefinition {
 
 	@Override
-	public ExtendedMetaTypeInformation getMetaTypeInformation(Bundle bundle);
+	public ExtendedAttributeDefinition[] getAttributeDefinitions(int filter);
+
+	public Map<String, String> getExtensionAttributes(String uri);
+
+	public Set<String> getExtensionUris();
 
 }
